+++ RELEASE_NOTES.md
@@ -1,5 +1,10 @@
 # Change log
 
+## in develop
+
+* Correctly attaches to docker stdout/stderr
+* Fixes `exec` command to correctly localize paths and handle return code when running in container
+
 ## 0.17.6 (2022-01-05)
 
 * Fixes `stderr()` function - previously it was returning the file for stdout